% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Ejercicio_Extra.R
\name{multiplicacion.dos.numeros}
\alias{multiplicacion.dos.numeros}
\title{Multiplicacion}
\usage{
multiplicacion.dos.numeros(numero_1, numero_2)
}
\arguments{
\item{numero_1}{numero que se quiere multiplicar.}

\item{numero_2}{numero que se quiere multiplicar.}
}
\description{
Esta funcion multiplica dos numeros.
}
\keyword{multiplicacion}
