% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Ejercicio_Extra.R
\name{potencia.dos}
\alias{potencia.dos}
\title{Potencia}
\usage{
potencia.dos(numero)
}
\arguments{
\item{numero}{numero que se quiere elevar a dos.}
}
\description{
Esta funcion devuelve la potencia de dos de un numero.
}
\keyword{potencia}
