% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Ejercicio_Extra.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Suma}
\usage{
suma.dos.numeros(numero_1, numero_2)
}
\arguments{
\item{numero_1}{numero que se quiere sumar.}

\item{numero_2}{numero que se quiere sumar.}
}
\description{
Esta funcion suma dos numeros.
}
\keyword{suma}
